# internal/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package dbgen

import (
	"github.com/jackc/pgx/v5/pgtype"
	decimal "github.com/shopspring/decimal"
)

type ApiKey struct {
	ID        string             `db:"id" json:"id"`
	KeyHash   string             `db:"key_hash" json:"key_hash"`
	Name      string             `db:"name" json:"name"`
	Scopes    []string           `db:"scopes" json:"scopes"`
	Active    bool               `db:"active" json:"active"`
	CreatedAt pgtype.Timestamptz `db:"created_at" json:"created_at"`
}

type Coupon struct {
	Code         string             `db:"code" json:"code"`
	DiscountType string             `db:"discount_type" json:"discount_type"`
	Value        decimal.Decimal    `db:"value" json:"value"`
	MinItems     int32              `db:"min_items" json:"min_items"`
	Description  string             `db:"description" json:"description"`
	Active       bool               `db:"active" json:"active"`
	CreatedAt    pgtype.Timestamptz `db:"created_at" json:"created_at"`
}

type Order struct {
	ID         string             `db:"id" json:"id"`
	Items      []byte             `db:"items" json:"items"`
	Total      decimal.Decimal    `db:"total" json:"total"`
	Discounts  decimal.Decimal    `db:"discounts" json:"discounts"`
	CouponCode string             `db:"coupon_code" json:"coupon_code"`
	CreatedAt  pgtype.Timestamptz `db:"created_at" json:"created_at"`
}

type Product struct {
	ID             string          `db:"id" json:"id"`
	Name           string          `db:"name" json:"name"`
	Price          decimal.Decimal `db:"price" json:"price"`
	Category       string          `db:"category" json:"category"`
	ImageThumbnail string          `db:"image_thumbnail" json:"image_thumbnail"`
	ImageMobile    string          `db:"image_mobile" json:"image_mobile"`
	ImageTablet    string          `db:"image_tablet" json:"image_tablet"`
	ImageDesktop   string          `db:"image_desktop" json:"image_desktop"`
}